Campaigns' Upgrade
Removed the chunks. The celery worker gets fed the entire list of leads associated with 1 account. 1 worker per 1 account's leads.
The chunking was causing multiple instances of the same account to behave simultaneously, which is not exactly intended for the human mimicking of random delay sending.
The delays were working fine for their own chunk but when a single accounts are used for multiple chunks, the purpose of delay wasn't fully achieved.