Improve email sending and error handling in warmup
Refactored email subject and body cleaning to handle encoding and non-breaking spaces.
Added specific handling for 'Username and Password not accepted' errors, disabling problematic accounts and notifying users.
Updated error handling to skip failed accounts instead of failing the entire campaign step.
Updated the templates text and making it ASCII only so that the problem is dealt with in the route.